Changed aip_id
Changed the way aip_id is defined for certain workflows to account for presence of qctools file, which has multiple dots in filename (ex. item.qctools.xml.gz). I got this solution from the wilds of the internet, and it seems to work after some testing, though I'm not sure if it's the most elegant solution.